cli e2e tests

+++ test/e2e/utils/TestUtils.js
@@ -1,6 +1,6 @@
-import fs from "fs";
-import { promisify } from "util";
-import cp from "child_process";
+import fs from 'fs';
+import { promisify } from 'util';
+import cp from 'child_process';
 
 const exec = promisify(cp.exec);
 
@@ -26,31 +26,56 @@ export function removeAnsiEscapeCodes(text) {
  * @param {string} directory The path to the directory the export files are located in. Default is the current directory "./".
  * @returns {Promise<void>}
  */
-export async function testExport(command, env, type, fileName, directory = "./", checkForMetadata = true) {
+export async function testExport(
+  command,
+  env,
+  type,
+  fileName,
+  directory = './',
+  checkForMetadata = true
+) {
   const { stdout } = await exec(command, env);
-  const regex = new RegExp(fileName ? fileName : (type ? `.*\\.${type}\\.(json|js|groovy)` : `.*\\.(json|js|groovy)`));
-  const filePaths = fs.readdirSync(directory).filter(n => regex.test(n)).map(n => `${directory}${directory.endsWith("/") ? '' : '/'}${n}`);
+  const regex = new RegExp(
+    fileName
+      ? fileName
+      : type
+      ? `.*\\.${type}\\.(json|js|groovy)`
+      : `.*\\.(json|js|groovy)`
+  );
+  const filePaths = fs
+    .readdirSync(directory)
+    .filter((n) => regex.test(n))
+    .map((n) => `${directory}${directory.endsWith('/') ? '' : '/'}${n}`);
   if (fileName) {
     expect(filePaths.length).toBe(1);
   } else {
     expect(filePaths.length >= 1).toBeTruthy();
   }
   expect(removeAnsiEscapeCodes(stdout)).toMatchSnapshot();
-  filePaths.forEach(path => {
-    if (path.endsWith("json")) {
-      const exportData = JSON.parse(fs.readFileSync(path, 'utf8'));
+  filePaths.forEach((path) => {
+    let deleteExportFile = true;
+    if (path.endsWith('json')) {
+      let exportData = {};
+      try {
+        exportData = JSON.parse(fs.readFileSync(path, 'utf8'));
+      } catch (error) {
+        deleteExportFile = false;
+        exportData = { path, error };
+      }
       if (checkForMetadata) {
         expect(exportData).toMatchSnapshot({
           meta: expect.any(Object),
-        });
+        }, path);
       } else {
-        expect(exportData).toMatchSnapshot();
+        expect(exportData).toMatchSnapshot(path);
       }
     } else {
       const data = fs.readFileSync(path, 'utf8');
-      expect(data).toMatchSnapshot();
+      expect(data).toMatchSnapshot(path);
     }
     //Delete export file
-    fs.unlinkSync(path);
+    if (deleteExportFile) fs.unlinkSync(path);
   });
 }
+
+export const testif = (condition) => (condition ? test : test.skip);
